+++ src/featureFlags/flags/debounceSwapQuote.ts
@@ -1,0 +1,7 @@
+import { BaseVariant, FeatureFlag, useBaseFlag } from '../index'
+
+export function useDebounceSwapQuoteFlag(): BaseVariant {
+  return useBaseFlag(FeatureFlag.debounceSwapQuote)
+}
+
+export { BaseVariant as DebounceSwapQuoteVariant }
