--- src/featureFlags/flags/featureFlags.ts
@@ -1,10 +1,0 @@
-/**
- * The value here must match the value in the statsig dashboard, if you plan to use statsig.
- */
-export enum FeatureFlag {
-  traceJsonRpc = 'traceJsonRpc',
-  permit2 = 'permit2',
-  fiatOnRampButtonOnSwap = 'fiat_on_ramp_button_on_swap_page',
-  detailsV2 = 'details_v2',
-  uraEnabled = 'ura_enabled',
-}
