+++ src/featureFlags/index.tsx
@@ -1,7 +1,18 @@
 import { atomWithStorage, useAtomValue, useUpdateAtom } from 'jotai/utils'
 import { createContext, ReactNode, useCallback, useContext } from 'react'
 import { useGate } from 'statsig-react'
-export { FeatureFlag } from './flags/featureFlags'
+
+/**
+ * The value here must match the value in the statsig dashboard, if you plan to use statsig.
+ */
+export enum FeatureFlag {
+  traceJsonRpc = 'traceJsonRpc',
+  permit2 = 'permit2',
+  fiatOnRampButtonOnSwap = 'fiat_on_ramp_button_on_swap_page',
+  detailsV2 = 'details_v2',
+  uraEnabled = 'ura_enabled',
+  debounceSwapQuote = 'debounce_swap_quote',
+}
 
 interface FeatureFlagsContextType {
   isLoaded: boolean
